Fase 5 (Automação): detecção de documento, deskew, perspectiva e melhoria automática
Novos ajustes de imagem em IImageService/SkiaImageService (seção 19):
- AdjustBrightnessContrast, AdjustGamma, AdjustSaturation, Sharpen, ReduceNoise,
  RemoveBackground — funções puras sobre PixelOps (extração de canal + desfoque de
  média separável compartilhado).
- CorrectPerspective: mapeia um quadrilátero arbitrário para um retângulo alinhado
  aos eixos via a homografia clássica de Heckbert (1989), com amostragem bilinear
  (seção 17).
- Rotate passa a aceitar ângulo double (antes int) — compatível com todos os
  chamadores existentes (literais 90/-90), necessário para o deskew em graus
  fracionários.

SCANOVA.Imaging.Detection (novo):
- ConvexHull (algoritmo de Andrew) + MinimumAreaRectangle ("rotating calipers",
  Toussaint 1983) — geometria clássica reutilizável.
- DocumentDetectionService (IDocumentDetectionService): segmenta documento/fundo
  por limiar de Otsu, decidindo a polaridade pela classe majoritária na borda da
  imagem (funciona com documento claro ou escuro); extrai o contorno externo,
  calcula o retângulo de área mínima e reporta quadrilátero + confiança (proporção
  de preenchimento) + ângulo de inclinação + tamanho físico estimado (seção 13/17).
- ProjectionProfileSkewEstimator: segundo estimador de inclinação (método de Postl,
  1986), independente de detecção de borda — funciona mesmo quando o documento
  preenche todo o quadro.

SCANOVA.Imaging.Enhancement (novo):
- BilevelDespeckle: remove pequenas manchas isoladas por componentes conectados
  (8-conectividade) após a binarização (seção 19).
- DocumentEnhancementService (IDocumentEnhancementService): orquestra os
  primitivos acima em um pipeline completo (perspectiva → inclinação → escala de
  cinza/remoção de fundo → redução de ruído → tom → nitidez → binarização +
  despeckle), com presets Leve/Normal/Forte traduzidos para SCANOVA.Core.Models.
  Apply() nunca modifica a imagem recebida (edição não destrutiva, seção 45).

SCANOVA.App:
- DocumentViewerPage/ViewModel: novos botões "Melhorar automaticamente" (aplica
  o preset Normal) e "Reverter para original".
- Dashboard "Melhorar documento" agora abre o seletor de arquivo e navega para o
  visualizador, em vez de um placeholder.

75 testes novos/atualizados em SCANOVA.Imaging.Tests (131 no total do
CrossPlatform.slnf), cobrindo os primitivos de ajuste, a homografia de
perspectiva, geometria (fecho convexo/retângulo mínimo), o estimador de
inclinação, detecção de documento (ambas as polaridades), despeckle e o
pipeline completo do motor de melhoria — incluindo um teste de ponta a ponta
que gira uma imagem sintética por um ângulo conhecido e confirma que o
deskew a endireita de volta.

+++ src/SCANOVA.Imaging/ImageProcessing/SkiaImageService.cs
@@ -12,7 +12,7 @@ namespace SCANOVA.Imaging.ImageProcessing;
 /// </summary>
 public sealed partial class SkiaImageService : IImageService
 {
-    public RasterImage Rotate(RasterImage image, int degrees)
+    public RasterImage Rotate(RasterImage image, double degrees)
     {
         var normalized = ((degrees % 360) + 360) % 360;
         if (normalized == 0)
@@ -31,7 +31,7 @@ public sealed partial class SkiaImageService : IImageService
         return RenderTransformed(src, newWidth, newHeight, canvas =>
         {
             canvas.Translate(newWidth / 2f, newHeight / 2f);
-            canvas.RotateDegrees(normalized);
+            canvas.RotateDegrees((float)normalized);
             canvas.Translate(-image.Width / 2f, -image.Height / 2f);
         }, image.HorizontalDpi, image.VerticalDpi);
     }
